refactor(abstract-utxo): split transaction handling tests into separate files
Improves organization by splitting large test file into separate focused
files for parseTransaction, explainTransaction, and verifyTransaction.

Issue: BTC-2732
